Move ExternalLinksComponent into the Show namespace
