+++ app/components/show/agreement_component.html.erb
@@ -2,7 +2,7 @@
 
 <div class="row">
   <div class="col">
-    <%= render ExternalLinksComponent.new(document: document) %>
+    <%= render Show::ExternalLinksComponent.new(document: document) %>
     <%= render Show::ControlsComponent.new(solr_document: document, manager: helpers.can?(:manage_item, presenter.cocina)) %>
   </div>
 </div>
