+++ app/components/show/external_links_component.html.erb
